Major changes to how Render Fix's custom shaders are loaded, and finally added 2D shader functionality.
Shaders are now directly compiled from source at startup, this not only makes development easier, but also allows us to "bake-in" certain features, such as dithering, without needing to ship tons of essentially copies of the same shader.

This means the shader module also needed updating, so I took the time to clean it up too and bring it to modern standards.